Move some OLA things out of the main ranking code. Delete test _test_scoring_with_elastic_search_conversions because we cant compare OLA results with the updated algorithm. Since the flips were updated, update a couple of tests.
